refactor: email templating using handlebars
