+++ packages/backend/package.json
@@ -20,8 +20,8 @@
     "@types/jest": "^29.5.6",
     "@types/lodash.partial": "^4.2.7",
     "@types/morgan": "^1.9.6",
-    "@types/nodemailer": "^6.4.13",
     "@types/node-fetch": "^2.6.4",
+    "@types/nodemailer": "^6.4.13",
     "@types/swagger-ui-express": "^4.1.4",
     "@types/uuid": "^9.0.4",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
@@ -48,6 +48,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-http-context": "^1.2.4",
+    "handlebars": "^4.7.8",
     "lodash.partial": "^4.2.1",
     "mongodb": "^6.1.0",
     "morgan": "^1.10.0",
